+++ changelog.md
@@ -2,8 +2,9 @@
 
 ## 0.23.2
 
-- add optional type generation for production to `esbuild` builder
-  ([#194](https://github.com/feltcoop/gro/pull/194))
+- generate types for production builds
+  ([#194](https://github.com/feltcoop/gro/pull/194),
+  [#195](https://github.com/feltcoop/gro/pull/195))
 
 ## 0.23.1